Add logical ("computed") axes
This was a substantial refactor, but it was very useful to clear up some of the core concepts involved.

The role of a Specification of an axis is now much better: it is a factory for an axis which `realize`s an axis during ManagedInstrument.__init__ time.

As a result, we do not need to have any metaprogramming in ManagedInstrument anymore and this has been removed.

All information relating to axis mocking are pulled into the axis specification which is cleaner, and there is a single sentinel driver "MockDriver" which signals that the instrument is being virtualized completely. Users can of course supply a separate driver that does whatever virtualization they like.

Another large change is that the UI construction happens over the instantiated axes and not the specification, which makes more sense. There is a primitive base class AxisView that organizes this for the time being.

Finally we get to logical axes:
Logical axes maintain references to the hardware axes and write to them/read from them after performing appropriate coordinate transforms. A suitable example has been added to this effect.